examples/metacall_faas/requirements.txt
fasthtml-extn